Fixed bugs, game over text, Mac and Linux support
I have fixed a couple of minor bugs in the program. Also, the text "Game
Over!" will now appear on the screen when you die, and you will be taken
back to the title screen after 1 second. I have also added .exe, .zip,
and .sh files for Windows, Mac, and Linux respectively for those who do
not want to run a .jar. The Mac and Linux versions will probably work,
but please note that I have not tested them.