Added enemy bullets, added player health
The player has health and dies when it goes down to 0.
The player displays an animation whenever it is hit.
Some of the enemies can also shoot bullets.
"Score" has been renamed to "Cash."
Hit animation now follows the enemies as they move
Minor tweaks to game mechanics and bug fixes