Added boss, tweaks to gameplay
There is now a boss that spawns at the 60 second mark, and you beat the
game if you beat it. The doubleshot has a decreased firing rate, and has
switched prices with the laser. The shop button has been remapped to the
F key. There are other minor bug fixes and tweaks in the gameplay.